mod#11/: viewmodel state 분리
