+++ app/src/main/java/com/sopt/now/compose/LoginState.kt
@@ -1,0 +1,6 @@
+package com.sopt.now.compose
+
+data class LoginState(
+    val isSuccess: Boolean,
+    val message: String
+)
