+++ app/src/main/java/com/sopt/now/compose/SignUpState.kt
@@ -1,0 +1,6 @@
+package com.sopt.now.compose
+
+data class SignUpState(
+    val isSuccess: Boolean,
+    val message: String
+)
