+++ app/src/main/java/com/sopt/now/compose/SignUpViewModel.kt
@@ -6,10 +6,6 @@ import androidx.lifecycle.ViewModel
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
-data class SignUpState(
-    val isSuccess: Boolean,
-    val message: String
-)
 
 class SignUpViewModel : ViewModel() {
     private val authService by lazy { ServicePool.authService }
