+++ app/src/main/java/com/sopt/now/compose/UserInfoState.kt
@@ -1,0 +1,9 @@
+package com.sopt.now.compose
+
+data class UserInfoState(
+    val isSuccess: Boolean,
+    val message: String,
+    val userId : String? = "",
+    val userNick : String? = "",
+    val userPhone : String? = ""
+)
